The main purpose of this update is to extend the storage device costing page.  That page is now functional and near its final form.  The main storage devices editing page was also updated to support the new costing fields recently added to the storagetypes table.
costhooks and cost_edit_fridge were extensively expanded to support storage device costing.  fridgehooks was modified to deal with the new costing fields on storage devices, and htmlgenerator was modified to respect the new fieldMap types 'currency' and 'price'.  hrmwidgets was modified to add a new widget editFormManager, and tweaks were made to currencySelector.  fridge_edit.tpl was modified to use editFormManager (which respects the new currency and price fieldMap types). fridge_top was modified to group the storage devices on the new 'Category' field.

currencyhelper.py was split from costhooks.py to avoid a circular import dependency.  costmodel.py gains a _toJSON method to avoid a problem with the debugging page being unable to encode the copy of the currency table which resides in the session.  input_defaults gains a record for amortizationstorageyears.  matcher got a json/no-op method for debugging purposes.  modelhooks was changed to always copy the currency table into new models as they are created.  typehelper gained the externally visible set hiddenTypesSet, to provide a sensible way to hide types like OUTDOORS from the UI.  ui_utils used to not import _()==translateString(), but it needs it for an error condition, so it was modified to do a delayed import.


git-svn-id: https://redmine.hindsight.psc.edu/svn/hermes/trunk@1792 c0c85e88-7b48-49d6-ad45-a8906df79eec