Added infrastructure to gather random cost-related events into lists and ultimately get them to the CostManager.
abstractbaseclasses gets CanOwn.applyToAll() to apply a user-supplied function to all owned things, and Costable gets getPendingCostEvents().  fridgetypes, trucktypes, and warehouse get implementations of those events; warehouse also gets infrastructure in the travel generator to append the pending cost list from each step to the travel generator.  dummycostmanager learns to clear pending cost events at the beginning of a costing interval.  legacycostmodel and model learn to ignore the extra tuple on the end of each tripJournal entry.  microcostmodel gets some improved ready-checking logic and other tweaks, but does not yet actually exploit the collected cost events.  

main gets a unicode fix.  runhooks gets a fix to logic which was breaking the run-model track when the previous run was of a modelId which has since been deleted.

git-svn-id: https://redmine.hindsight.psc.edu/svn/hermes/trunk@1961 c0c85e88-7b48-49d6-ad45-a8906df79eec