This checkin adds basic support for limited warm storage.  By default things should work close to identically to how it worked before.  You need to turn on the option "limitedroomtemp" in the input kvp file for this to be active.
There are a few things that are still utterly absurd and brittle as to how they are coded 
currently with me taking the attitude that it was better to get something working at all than 
to try to get them to work properly initially.  If we want this feature of hermes to be fully
supported then these should be fixed sooner than later.  The biggest was recognizing the 
"outdoors" "fridge" and knowing that it should be treated special in different places than
it was in the old code and to do that I looked for a volume of 1000000000000 cc.






git-svn-id: https://redmine.hindsight.psc.edu/svn/hermes/trunk@2294 c0c85e88-7b48-49d6-ad45-a8906df79eec