Some changes to make the installation process more flexible.
Site Info now has a hierarchy of looking for a system level kvp that is in the main HERMES directory and then a user level one that is in the local directory, and then one can specify as an initialization argument a specific config file.

install_hermes.py now uses the site_info more heavily rather than having hard coded directories.

Now working on the windows installation scripts.py