Added a BufferStockFraction to allow for intimate control of the buffer stock of each location in the model.py
There is a new input paramter called bufferstockfraction that will set the default buffer stock fraction for the whole model.py

You can add a column to the stores file that is called BufferStockFraction that will override the default for an individual location.

The default is set in the input_default.csv file as 0.25.

This has been incorporated in the the database as well, and so an alembic upgrade head needs to be executed to use this feature.