Added a parameter that allows one to define a fraction of doses per vial that one has to get to to open a vial.
By default, this is off.

The input parameter is openvialdenyfraction, and if it is less than 0, it is off, if it is greater than 1.00 then everything thing that is not fully using a vial will be denied.  Anything in between, if the last vial to be used has a population demand that is less than (non inclusive) floor(doses per vial*fraction)
then the vial will not be used.

This can only be set globally, and I have tested it, but if used, one should be careful to make sure the results look like they make sense.