Finalized some new widgets in hrmwidgets.tpl
simpleTypeSelectField provides a select box for a particular model and type class (e.g. fridges) that then has a popup next to it that gives detailed info about the type as you move through the list.

typeInfoPopup is a unformated popup that gives you detailed information about a type instance

typeIntoDialog is a dialog box similar to the ones that come up with info

typeInfoButtonAndDialog provides the info button with an associated div to put the dialog in so that this can be placed anywhere in the gui

simpleStorageDeviceTable is an editable table that allows you to change the counts of something.  Need to make it generic for counts of any types