Lots of changes to the experiments pages
I have homogenized all but the vaccine intro experiment to have the working page while the model is being modified.

The adding loops work.

fixed a couple minor bugs in slideshow.

Also added the ability to hide the button row at the top of the slide show and put a border around it.