Fix issue 68
The first time you go to the "specify vaccine dose schedule" page (or any time you enter the page with no demand entries in the db), in addition to adding one "0" entry to each row (each vaccine gets a zero entry in one population type) we now add an additional "0" entry to each column (each population type gets a zero entry in one vaccine type). This is enough to create a full table.